backend/requirements.txt
# Core AI & ML Libraries
chromadb==0.4.18
ollama==0.1.7
groq==0.4.1
openai-whisper==20231117
sentence-transformers==2.2.2
transformers==4.36.0
torch>=2.7.0

# Vector & Embedding Libraries
faiss-cpu==1.7.4
numpy>=2.1.0
scikit-learn==1.3.0

# Database & Storage
redis==5.0.1
firebase-admin==6.2.0
pymongo==4.6.0

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
httpx==0.25.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Speech & Audio
pyaudio==0.2.11
pydub==0.25.1
festival==0.1.0
espeak==0.6.0

# Language Processing
googletrans==4.0.0
langdetect==1.0.9
spacy==3.7.2
nltk==3.8.1

# Utilities & Monitoring
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0
schedule==1.2.0
watchdog==3.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# n8n Integration
requests==2.31.0
celery==5.3.4
kombu==5.3.4
